+++ src/include/searchresults.hpp
@@ -1,5 +1,6 @@
+#include <cstdint>
 #include <vector>
 struct SearchResults {
-  std::vector<int> approximateNN;
-  std::vector<int> visited;
+  std::vector<int64_t> approximateNN;
+  std::vector<int64_t> visited;
 };
